Mild refactor of native_functions.yaml dispatch parsing
This refactor is no longer necessary for ufunc codegen, as I changed
the format of ufuncs to not directly be inserted into the 'dispatch'
key, but I think the refactored code here is better.  The basic concept
is to directly construct BackendMetadata as we are parsing entries of
the dispatch dictionary, rather than post facto creating them later.
This centralizes the compute and means that the creation of the backend index
is just a simple reindexing by operator name (nothing nontrivial).

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 7cdb2f3ed1b6e796272bec9e7cc8ad8f9a0225d7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66109